requirements.txt
llama-cpp-python==0.2.23
picovoice==3.0.2
python-dotenv==1.0.0
pydantic==2.6.1 